wBuildup Scene procedural update
working on the projectile and weapons system. this is a procedural update that organizes a few of my scripts into proper folders. also the fireball projectile prefab has been created and tested. it works using the proper interfaces / scripts.

nothing breaks as far as i can tell. only working on new objects.